+++ src/pvp/gui.ts
@@ -1,8 +1,8 @@
 import { assert } from '../misc/assert'
 import { prestart } from '../loading-stages'
-import { PvpTeam } from './pvp'
 import { Opts } from '../options'
 import { COLOR, wrapColor } from '../misc/wrap-color'
+import { MultiParty } from '../party/party'
 
 import './combatant-gui-fix'
 
@@ -28,16 +28,16 @@ function injectIntoPvpUpperGui(clazz: sc.CombatUpperHud.CONTENT_GUI.PVP_CONSTRUC
     let x!: number
     let renderer!: ig.GuiRenderer
 
-    function drawTeamName(team: PvpTeam) {
-        const text = new ig.TextBlock(sc.fontsystem.smallFont, team.name, {
+    function drawTeamName(party: MultiParty) {
+        const text = new ig.TextBlock(sc.fontsystem.smallFont, party.title, {
             maxWidth: undefined,
         })
         renderer.addText(text, x, 0)
         x += text.size.x
     }
 
-    function drawTeamHeads(this: sc.CombatUpperHud.CONTENT_GUI.PVP, team: PvpTeam, left: boolean) {
-        const heads: number[] = team.players.map(player => player.getHeadIdx())
+    function drawTeamHeads(this: sc.CombatUpperHud.CONTENT_GUI.PVP, party: MultiParty, left: boolean) {
+        const heads: number[] = multi.server.party.getPartyCombatants(party).map(player => player.getHeadIdx())
         if (left) x += (heads.length - 1) * 16
         this._renderHeads(renderer, x + (left ? 24 : 0), left, heads)
         if (left) x += 16
@@ -45,11 +45,11 @@ function injectIntoPvpUpperGui(clazz: sc.CombatUpperHud.CONTENT_GUI.PVP_CONSTRUC
         x += 8
     }
 
-    function drawTeamPoints(this: sc.CombatUpperHud.CONTENT_GUI.PVP, team: PvpTeam, left: boolean) {
+    function drawTeamPoints(this: sc.CombatUpperHud.CONTENT_GUI.PVP, party: MultiParty, left: boolean) {
         if (sc.pvp.winPoints != 1) {
             x += 4
             if (left) x += (sc.pvp.winPoints - 2) * 5
-            this._renderPoints(renderer, x, left ? -1 : 1, sc.pvp.winPoints, sc.pvp.points[team.party]!, 0)
+            this._renderPoints(renderer, x, left ? -1 : 1, sc.pvp.winPoints, sc.pvp.points[party.combatantParty]!, 0)
 
             if (left) x += 5
             else x += sc.pvp.winPoints * 5
@@ -62,15 +62,15 @@ function injectIntoPvpUpperGui(clazz: sc.CombatUpperHud.CONTENT_GUI.PVP_CONSTRUC
         x += 16
     }
 
-    function drawTeam(this: sc.CombatUpperHud.CONTENT_GUI.PVP, team: PvpTeam, left: boolean) {
+    function drawTeam(this: sc.CombatUpperHud.CONTENT_GUI.PVP, party: MultiParty, left: boolean) {
         if (left) {
-            drawTeamName(team)
-            drawTeamHeads.call(this, team, left)
-            drawTeamPoints.call(this, team, left)
+            drawTeamName(party)
+            drawTeamHeads.call(this, party, left)
+            drawTeamPoints.call(this, party, left)
         } else {
-            drawTeamPoints.call(this, team, left)
-            drawTeamHeads.call(this, team, left)
-            drawTeamName(team)
+            drawTeamPoints.call(this, party, left)
+            drawTeamHeads.call(this, party, left)
+            drawTeamName(party)
         }
     }
 
@@ -81,11 +81,11 @@ function injectIntoPvpUpperGui(clazz: sc.CombatUpperHud.CONTENT_GUI.PVP_CONSTRUC
             x = 0
             renderer = renderer1
 
-            drawTeam.call(this, sc.pvp.teams[0], sc.pvp.teams.length == 2)
+            drawTeam.call(this, sc.pvp.parties[0], sc.pvp.parties.length == 2)
 
-            for (let i = 1; i < sc.pvp.teams.length; i++) {
+            for (let i = 1; i < sc.pvp.parties.length; i++) {
                 drawVsGfx.call(this)
-                drawTeam.call(this, sc.pvp.teams[i], false)
+                drawTeam.call(this, sc.pvp.parties[i], false)
             }
 
             this.setSize(x + 5, 20)